requirements.txt
# Core & Web Framework
asyncio-mqtt>=0.11.0
aiofiles>=0.8.0
aiohttp>=3.8.0
aioredis>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
httpx>=0.25.0
requests>=2.31.0

# LangChain & LlamaIndex (AI Orchestration)
langchain>=0.0.325
langchain-experimental>=0.0.40
langchain-community>=0.0.10
llama-index>=0.8.60
langchain-google-genai>=0.0.8
langchain-openai>=0.0.5
langchain-huggingface>=0.0.3

# Vector Databases & Search
faiss-cpu>=1.7.4
weaviate-client>=3.25.3
qdrant-client>=1.6.4
pymilvus>=2.3.2

# BentoML & MLflow (Model Serving & Tracking)
bentoml>=1.1.10
mlflow>=2.8.1

# DVC (Data Version Control)
dvc>=3.27.0
dvc-s3>=2.23.0

# Workflow Orchestration
celery>=5.3.4
celery[redis]>=5.3.4
prefect>=2.14.10
dagster>=1.5.10

# Monitoring & Observability
psutil>=5.9.6
prometheus-client>=0.18.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-celery>=0.42b0
opentelemetry-exporter-prometheus>=0.42b0
sentry-sdk[fastapi]>=1.38.0

# Security & Code Quality
bandit>=1.7.5
safety>=2.3.5
cryptography>=41.0.7
bcrypt>=4.0.1
PyJWT>=2.8.0

# Development Tools
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0
black>=23.11.0
ruff>=0.1.6
isort>=5.12.0
pre-commit>=3.5.0
mypy>=1.7.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
click>=8.1.7
tqdm>=4.66.1
pyyaml>=6.0.1
python-multipart>=0.0.6
jinja2>=3.1.2

# AI Models & Transformers
torch>=2.1.1
torchvision>=0.16.1
transformers>=4.35.2
sentence-transformers>=2.2.2
accelerate>=0.24.1
bitsandbytes>=0.41.1
peft>=0.6.2
trl>=0.7.4
datasets>=2.14.7
tokenizers>=0.14.1

# Model Providers (OpenAI, Anthropic, etc.)
openai>=1.3.7
anthropic>=0.7.2
google-generativeai>=0.3.1

# Ollama & Local Models
ollama>=0.1.7

# Embedding & Vector Search
sentencepiece>=0.1.99
einops>=0.7.0
safetensors>=0.4.0

# Data Processing
nltk>=3.8.1
spacy>=3.7.2

# Asynchronous Performance
uvloop>=0.19.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.4.8
mkdocs-swagger-ui-tag>=0.6.2

# Testing & Mocking
fakeredis>=2.20.1
hypothesis>=6.88.1
factory-boy>=3.3.0

# Jupyter & Interactive
ipykernel>=6.26.0
notebook>=7.0.6

# AI Agent Frameworks
metagpt>=0.8.1
autogpt>=0.5.0
babyagi>=0.2.0
autogen>=0.2.0

# Model Serving
vllm>=0.2.5

# WebRTC
aiortc>=1.6.0

# Chinese Models
moonshot>=0.1.0
minimax>=0.1.0
